Fixed comments to docstrings and added dimension test case
